Back To Dev

+++ OneCoin/Server/src/main/java/OneCoin/Server/deposit/entity/Deposit.java
@@ -8,6 +8,7 @@ import lombok.NoArgsConstructor;
 import lombok.Setter;
 
 import javax.persistence.*;
+import java.math.BigDecimal;
 
 @AllArgsConstructor
 @NoArgsConstructor
@@ -24,7 +25,7 @@ public class Deposit extends CreatedOnlyAuditable {
     private Balance balance;
 
     @Column(nullable = false)
-    private Long remainingBalance;
+    private BigDecimal remainingBalance;
 
     @Column(nullable = false)
     private Long depositAmount;
